Rename rest -> api.
